Added changed rotation to server and client.
The server now parses the rotation matrix from the avatars and sends the message to the other clients to update the ghost avatars rotation matrix.

Also added the backwards and yaw movement to avatars.

Changed the ghost avatar model to the dolphin model. Ghost avatars are now red avatars. After we create our own models and add them to the game, I will need to parse the models that each client choose and send those messages over to other clients.